CAMEL-9799: Switched to use ResourceEndpoint
