+++ components/camel-everit-json-schema/src/main/docs/json-validator-component.adoc
@@ -1,9 +1,7 @@
 == JSON Schema Validator Component
 === Everit Json Schema Validator Component
-*Available as of Camel version *
 
-
-*Available as of Camel version 2.20*
+*Available as of Camel version 2.21*
 
 The JSON Schema Validator component performs bean validation of the message body
 agains JSON Schemas using the Everit.org JSON Schema library
@@ -57,14 +55,15 @@ with the following path and query parameters:
 [width="100%",cols="2,5,^1,2",options="header"]
 |===
 | Name | Description | Default | Type
-| *resourceUri* | *Required* URL to a local resource on the classpath or a reference to lookup a bean in the Registry or a full URL to a remote resource or resource on the file system which contains the JSON Schema to validate against. |  | String
+| *resourceUri* | *Required* Path to the resource. You can prefix with: classpath file http ref or bean. classpath file and http loads the resource using these protocols (classpath is default). ref will lookup the resource in the registry. bean will call a method on a bean to be used as the resource. For bean you can specify the method name after dot eg bean:myBean.myMethod. |  | String
 |===
 
-==== Query Parameters (6 parameters):
+==== Query Parameters (7 parameters):
 
 [width="100%",cols="2,5,^1,2",options="header"]
 |===
 | Name | Description | Default | Type
+| *contentCache* (producer) | Sets whether to use resource content cache or not | false | boolean
 | *failOnNullBody* (producer) | Whether to fail if no body exists. | true | boolean
 | *failOnNullHeader* (producer) | Whether to fail if no header exists when validating against a header. | true | boolean
 | *headerName* (producer) | To validate against a header instead of the message body. |  | String
