+++ components/camel-everit-json-schema/src/test/resources/org/apache/camel/component/everit/jsonschema/schema.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "http://json-schema.org/draft-04/schema#", 
+  "$schema": "http://json-schema.org/draft-06/schema#", 
   "definitions": {}, 
   "id": "http://example.com/example.json", 
   "properties": {
